Make two deviations from txrep (SEP must be fixed) to work around issues:
Zero-length vectors had no txrep representation.  Now we represent them
as the single digit "0" (as opposed to a single zero-value byte, which
continues to be "00").  We render it with the comment bytes (i.e., "0 bytes")
so that this is less intuitive.

For pointers, default to true if there is any valuea assigned to the pointer
and _present has not been specified.  This is because it makes descriptions
easier to read to omit the _present field (even though it is still output).